Implemented Player shooting projectiles with a delay, projectiles do currently not spawn infront of the gun. We probably need to add some predetermined offset for where the gun should be relative to the sprite, taking rotation into consideration as we do right now, as well as the sprite size.
Bounding boxes in the gameModel need to play a part in this I believe and we probably need to have our entities bounding boxes be based on how we decide to scale or sprites. So communicating this View -> Controller -> Model at game initialization similar to how we do with where the center of the screen is. (This is also needed for collisions in the model to be consistent with when sprites are colliding in the view.)