Simplified enemy design.
Adding a FacePlayer component to an entity simply results in the entity rotating towards facing the player at each update.
This component may be used for rotating stationary turrets simply by assigning a GameObject this component and a FireAtPlayer component, excluding a physicscomponent.

Adding a FireAtPlayer component which takes a value for damage, firing range and the cooldown inbetween shots.
By attaching FacePlayer, FireAtPlayer and a physics component to an enemy we get a chaser like enemy which fires at the player.
But we could for instance have an object which moves in a line like an asteroid firing at the player by changing the firing angle to be towards the player and not the current forward direction of the entity.